+++ packages/universal-xml-plugin/test/unit/plugin.spec.js
@@ -1,6 +1,12 @@
 import {UniversalXMLPlugin} from '../..';
 import BaseDriver from 'appium/driver';
-import {XML_IOS, XML_ANDROID, XML_IOS_TRANSFORMED, XML_ANDROID_TRANSFORMED} from '../fixtures';
+import {
+  XML_IOS,
+  XML_ANDROID,
+  XML_IOS_TRANSFORMED,
+  XML_ANDROID_TRANSFORMED,
+  XML_WEBVIEW,
+} from '../fixtures';
 import {runQuery, getNodeAttrVal} from '../../lib/xpath';
 
 describe('UniversalXMLPlugin', function () {
@@ -9,11 +15,13 @@ describe('UniversalXMLPlugin', function () {
   describe('getPageSource', function () {
     const driver = new BaseDriver();
     it('should transform page source for ios', async function () {
+      driver.getCurrentContext = () => 'NATIVE_APP';
       next = driver.getPageSource = () => XML_IOS;
       driver.caps = {platformName: 'iOS'};
       await p.getPageSource(next, driver).should.eventually.eql(XML_IOS_TRANSFORMED);
     });
     it('should transform page source for android', async function () {
+      driver.getCurrentContext = () => 'NATIVE_APP';
       next = driver.getPageSource = () => XML_ANDROID;
       driver.caps = {platformName: 'Android'};
       driver.opts = {appPackage: 'io.cloudgrey.the_app'};
@@ -24,6 +32,7 @@ describe('UniversalXMLPlugin', function () {
   describe('findElement(s)', function () {
     const driver = new BaseDriver();
     it('should turn an xpath query into another query run on the original ios source', async function () {
+      driver.getCurrentContext = () => 'NATIVE_APP';
       next = driver.getPageSource = () => XML_IOS;
       driver.caps = {platformName: 'iOS'};
       // mock out the findElement function to just return an xml node from the fixture
@@ -37,6 +46,7 @@ describe('UniversalXMLPlugin', function () {
     });
 
     it('should turn an xpath query into another query run on the original android source', async function () {
+      driver.getCurrentContext = () => 'NATIVE_APP';
       next = driver.getPageSource = () => XML_ANDROID;
       driver.caps = {platformName: 'Android'};
       driver.opts = {appPackage: 'io.cloudgrey.the_app'};
@@ -47,6 +57,21 @@ describe('UniversalXMLPlugin', function () {
       const node = await p.findElement(next, driver, 'xpath', '//TextInput[@axId="username"]');
       getNodeAttrVal(node, 'content-desc').should.eql('username');
       node.nodeName.should.eql('android.widget.EditText');
+    });
+
+    it('should not modify the xpath query and proxy the call to underlying driver', async function () {
+      driver.getCurrentContext = () => 'WEB_VIEW';
+      driver.findElement = () => {};
+      driver.caps = {platformName: 'Android'};
+      driver.opts = {appPackage: 'io.cloudgrey.the_app'};
+      const selector = '//div[@id="section-1"]';
+      next = () => {
+        const nodes = runQuery(selector, XML_WEBVIEW);
+        return nodes[0];
+      };
+      const node = await p.findElement(next, driver, 'xpath', selector);
+      getNodeAttrVal(node, 'id').should.eql('section-1');
+      node.nodeName.should.eql('div');
     });
   });
 });